requirements.txt
asgiref==3.8.1
boto3==1.35.83
botocore==1.35.83
certifi==2024.8.30
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
coverage==7.6.8
cryptography==44.0.0
distlib==0.3.9
dj-database-url==2.3.0
Django==5.1.4
django-admin-volt==1.0.14
django-allauth==65.2.0
django-ckeditor==6.7.0
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-js-asset==3.0.1
django-storages==1.14.4
djangorestframework==3.15.2
factory_boy==3.3.1
Faker==33.1.0
filelock==3.16.1
h11==0.14.0
identify==2.6.3
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
jmespath==1.0.1
nodeenv==1.9.1
oauthlib==3.2.2
packaging==24.2
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.0.1
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.10.1
pytest==8.3.3
pytest-django==4.9.0
pytest-factoryboy==2.7.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
qrcode==8.0
requests==2.32.3
requests-oauthlib==2.0.0
s3transfer==0.10.4
six==1.17.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
uvicorn==0.32.1
virtualenv==20.28.0
whitenoise==6.8.2
